apps/ai-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.10
python-multipart>=0.0.18,<1.0.0
python-dotenv>=1.0.1,<2.0.0
openai>=1.0.0,<2.0.0
pinecone>=7.0.0,<8.0.0
